+++ Modules/Microsoft365DSC/Examples/Resources/AADTokenLifetimePolicy/1-ConfigureAADTokenLifetimePolicy.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -20,7 +20,7 @@ Configuration Example
             Definition            = @('{"TokenIssuancePolicy":{"Version": 1,"SigningAlgorithm": "http://www.w3.org/2000/09/xmldsig#rsa-sha1","TokenResponseSigningPolicy": "TokenOnly","SamlTokenVersion": "2.0"}}')
             IsOrganizationDefault = $false
             Ensure                = "Present"
-            Credential            = $credsGlobalAdmin
+            Credential            = $Credscredential
         }
     }
 }