+++ Modules/Microsoft365DSC/Examples/Resources/EXOGlobalAddressList/1-ConfigureGlobalAddressList.ps1
@@ -9,7 +9,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -22,7 +22,7 @@ Configuration Example
             ConditionalDepartment        = "Human Resources"
             ConditionalStateOrProvince   = "Washington"
             Ensure                       = "Present"
-            Credential                   = $credsGlobalAdmin
+            Credential                   = $Credscredential
         }
     }
 }