+++ Modules/Microsoft365DSC/Examples/Resources/IntuneDeviceCompliancePolicyAndroidDeviceOwner/2-RemoveDeviceCompliancePolicyAndroidDeviceOwner.ps1
@@ -7,7 +7,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -15,9 +15,9 @@ Configuration Example
     {
         IntuneDeviceCompliancePolicyAndroidDeviceOwner 'RemoveAndroidDeviceCompliancePolicyOwner'
         {
-            DisplayName = 'DeviceOwner'
+            DisplayName = 'DeviceOwnerPolicy'
             Ensure      = 'Absent'
-            Credential  = $credsGlobalAdmin
+            Credential  = $Credscredential
         }
     }
 }