+++ Modules/Microsoft365DSC/Examples/Resources/IntuneWifiConfigurationPolicyAndroidEnterpriseDeviceOwner/1-ConfigureIntuneWifiConfigurationPolicyAndroidEnterpriseDeviceOwner.ps1
@@ -9,7 +9,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -33,7 +33,7 @@ Configuration Example
             ProxySettings                  = 'none'
             Ssid                           = 'MySSID - 3'
             Ensure                         = 'Present'
-            Credential                     = $credsGlobalAdmin
+            Credential                     = $Credscredential
         }
     }
 }