+++ Modules/Microsoft365DSC/Examples/Resources/TeamsGuestMessagingConfiguration/1-ConfigureTeamsGuestMessaging.ps1
@@ -7,7 +7,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -24,7 +24,7 @@ Configuration Example
             AllowUserDeleteMessage = $False
             AllowUserEditMessage   = $True
             GiphyRatingType        = "Moderate"
-            Credential             = $credsglobaladmin
+            Credential             = $Credscredential
         }
     }
 }