+++ Tests/Integration/Microsoft365DSC/M365DSCIntegration.INTUNE.Tests.ps1
@@ -339,7 +339,7 @@
                     ManagedEmailProfileRequired                 = $True
                     Ensure                                      = 'Present'
                     Credential                                  = $Credscredential
-
+        
                 }
                 IntuneDeviceCompliancePolicyiOs 'RemoveDeviceCompliancePolicyiOS'
                 {
@@ -461,7 +461,7 @@
                                         {
                                             Name = 'hosted_app'
                                         }
-
+        
                                         MSFT_IntuneGroupPolicyDefinitionValuePresentationValueKeyValuePair
                                         {
                                             Name = 'user_script'
@@ -495,7 +495,7 @@
                                     Id                          = '14c48993-35af-4b77-a4f8-12de917b1bb9'
                                     odataType                   = '#microsoft.graph.groupPolicyPresentationValueDecimal'
                                 }
-
+        
                                 MSFT_IntuneGroupPolicyDefinitionValuePresentationValue
                                 {
                                     presentationDefinitionId    = '98998e7f-cc2a-4d96-8c47-35dd4b2ce56b'
@@ -504,7 +504,7 @@
                                     Id                          = '4d654df9-6826-470f-af4e-d37491663c76'
                                     odataType                   = '#microsoft.graph.groupPolicyPresentationValueDecimal'
                                 }
-
+        
                                 MSFT_IntuneGroupPolicyDefinitionValuePresentationValue
                                 {
                                     presentationDefinitionId    = '6900e752-4bc3-463b-9fc8-36d78c77bc3e'
@@ -2681,12 +2681,5 @@
     }
 
     # Compile and deploy configuration
-    try
-    {
-        Master -ConfigurationData $ConfigurationData -Credscredential $Credential
-        Start-DscConfiguration Master -Wait -Force -Verbose -ErrorAction Stop
-    }
-    catch
-    {
-        throw $_
-    }
+    Master -ConfigurationData $ConfigurationData -Credscredential $Credential
+    Start-DscConfiguration Master -Wait -Force -Verbose