+++ CHANGELOG.md
@@ -7,7 +7,7 @@
     FIXES [#3904](https://github.com/microsoft/Microsoft365DSC/issues/3904)
 * IntuneAntivirusPolicyWindows10SettingCatalog
   * Fix issue with Set-TargetResource when retieving a policy from displayName
-    FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/3904)
+    FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/4003)
 * IntuneWindowsUpdateForBusinessRingUpdateProfileWindows10
   * Fix parameter name in assignment cmdlet
     FIXES [#4007](https://github.com/microsoft/Microsoft365DSC/issues/4007)