+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   * Fix issue with assignments
     FIXES [#3904](https://github.com/microsoft/Microsoft365DSC/issues/3904)
 * IntuneAntivirusPolicyWindows10SettingCatalog
-  * Fix issue with Set-TargetResource when retieving a policy from displayName
+  * Fix issue with Set-TargetResource when retrieving a policy from displayName
     FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/4003)
 * IntuneWindowsUpdateForBusinessRingUpdateProfileWindows10
   * Fix parameter name in assignment cmdlet