+++ CHANGELOG.md
@@ -6,8 +6,11 @@
   * Fix issue with assignments
     FIXES [#3904](https://github.com/microsoft/Microsoft365DSC/issues/3904)
 * IntuneAntivirusPolicyWindows10SettingCatalog
-  * Fix issue with Set-TargetResource when retieving a policy from displayName
-    FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/3904)
+  * Fix issue with Set-TargetResource when retrieving a policy from displayName
+    FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/4003)
+* IntuneWindowsUpdateForBusinessRingUpdateProfileWindows10
+  * Fix parameter name in assignment cmdlet
+    FIXES [#4007](https://github.com/microsoft/Microsoft365DSC/issues/4007)
 * DEPENDENCIES
   * Updated Microsoft.Graph to version 2.11.0.
   * Updated MSCloudLoginAssistant to version 1.1.3.